+++ dwds/CHANGELOG.md
@@ -14,6 +14,7 @@
 - Fix chrome detection in iPhone emulation mode in chrome or edge browsers.
 - Reliably find unused port for extension backend http service.
 - Ignore offset / count parameters in getObject if the object has no length
+- Implement `lookupResolvedPackageUris` and `lookupPackageUris` vm service API.
 
 ## 11.4.0